+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased:
 
+- Watch assets dir for changes in watch mode
 - Added `refresh-templates` task to update to latest templates
 - Social sharing (preview for Facebook, Twitter, LinkedIn, etc.)
 - Move metadata to post files and improve caching